+++ src/pages/payments/register/complete/index.ts
