+++ src/pages/payments/register/form/constant.ts
