+++ src/pages/payments/register/form/hooks/usePassword.ts
