+++ src/pages/payments/register/form/index.ts
@@ -1,3 +1,3 @@
-export { Payments } from './Payments';
+export { Form } from './Form';
 export type { ExpirationDate } from './types';
 export { BRAND_NUMBER } from './constant';
