+++ src/pages/payments/register/form/validator.ts
